server/py/pyproject.toml
[tool.poetry]
name = "tabletalkpoker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
pyserial = "^3.5"
flask-sock = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
